Final brain data frame ready for the analyses
I still had no time today to perform all the analyses that we both previously performed. I reuploaded this code in Rmd format. I think we'll be able to perform similar analyses to the previous Juan did. The final data frame is 500MB. In this case, the cloud force for the analyses will be needed.